[ARM] msm: usb: gadget: Do not suspend the phy when VBUS is off
When vbus is off driver tries to put the phy in suspend mode. As we
dont have complete infrastructure for LPM yet; Dont put phy in LPM
LPM as we will not be able to wakeup from the same. LPM support will be
implemented in a separate patch

Patch provided by Qualcomm

Signed-off-by: Mike Lockwood <lockwood@android.com>