[ARM] msm: kgsl: Map all memory allocations through the MMU
Map all memory allocations through the MMU and set it up to fault
if an address does not lie within its virtual address range. The
only allocations that are not mapped are the allocations for MMU
page table and a dummyspace allocation whose address is written
to the REG_MH_MMU_TRAN_ERROR register.

Change-Id: I92f93d87a9d4d2551d58520d46d820f164b150dc
Signed-off-by: Simon Wilson <swilson@quicinc.com>
Signed-off-by: Dima Zavin <dima@android.com>