sensors: board-supersonic
this fixed the proximity sensor and enabled the rest of the sensors to work \o/. anyhow the board file was the real winner here. but i kanged some other things along the way.

msm_hsusb_8x50_phy_reset commented out for compile reasons and because i havent add the usb shit yet.

Change-Id: I468ff8fc160ce39ca6a1ecc6067dc9ce36475a72

Signed-off-by: Brandon Berhent <bbedward@gmail.com>